requirements-dev.txt
asgiref==3.2.10
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-cors-headers==3.7.0
djangorestframework==3.11.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.0
Pillow==8.3.1
plotly==4.14.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
todo==0.1
urllib3==1.26.2
Werkzeug==1.0.1
gunicorn==20.0.4
psycopg2-binary==2.8.6